hardware/plant-monitor/requirements.txt
growhat
paho-mqtt
ltr559